refactor: move resolveCompilerMode to compiler-mode.ts and remove driver-model.ts
